requirements.txt
numpy~=1.16.0
six~=1.15.0
tensorflow~=2.2.0
tensorflow-gpu~=2.2.0
scipy~=1.4.1
scikit-learn~=0.23.1
opencv-python~=4.3.0.36
matplotlib~=3.2.1
Pillow~=6.2.2
PyYAML~=3.13
setuptools~=41.2.0
pandas~=1.0.5
scikit-image~=0.17.2
h5py~=2.10.0
future~=0.18.2
nltk~=3.5
gensim~=3.8.3
lxml~=4.5.2
networkx~=2.4
PyQt5~=5.15.0
pathlib~=1.0.1
httplib2~=0.18.1
google-api-python-client~=1.9.3
oauth2client~=4.1.3
typing~=3.7.4.1
sentencepiece~=0.1.91
dataclasses~=0.6
pycocotools~=2.0
contextlib2~=0.6.0.post1
pip~=20.1.1
wheel~=0.34.2
py~=1.9.0
Cython~=0.29.14
imageio~=2.9.0
pyformulas~=0.2.11
pytesseract~=0.3.4
textract~=1.6.3
spacy~=2.3.2
Keras~=2.4.3
Flask~=1.1.2
Werkzeug~=1.0.1
lightgbm~=2.3.1
joblib~=0.16.0
pyenchant~=3.1.1
python-dateutil~=2.8.1
tqdm~=4.47.0
imutils~=0.5.3
whitening~=0.1
torch~=1.6.0
torchaudio~=0.6.0
pdf2image~=1.13.1
datefinder~=0.7.1
requests~=2.24.0